♻️ refactor:  ディレクトリ名変更
Issue #285
