+++ lib/presentation/components/error_view.dart
