+++ lib/presentation/components/search_app_bar.dart
