+++ lib/presentation/pages/repo/repo_index_page.dart
@@ -6,7 +6,7 @@ import 'package:flutter/material.dart';
 
 import '../../../utils/logger.dart';
 import '../../../utils/routing/router.dart';
-import '../../widgets/search_app_bar.dart';
+import '../../components/search_app_bar.dart';
 import 'components/repo_list_view.dart';
 import 'components/search_repos_query.dart';
 import 'components/search_repos_sort.dart';
