+++ test/test_utils/test_agent.dart
@@ -21,9 +21,8 @@ import 'package:github_search/infrastructure/hive/hive.dart';
 import 'package:github_search/infrastructure/isar/isar.dart';
 import 'package:github_search/infrastructure/isar/query_history/query_history_repository.dart';
 import 'package:github_search/presentation/app.dart';
+import 'package:github_search/presentation/components/cached_circle_avatar.dart';
 import 'package:github_search/presentation/pages/repo/components/readme_markdown.dart';
-import 'package:github_search/presentation/pages/repo/components/search_repos_query.dart';
-import 'package:github_search/presentation/widgets/cached_circle_avatar.dart';
 import 'package:github_search/utils/localizations/strings.g.dart';
 import 'package:hive_flutter/hive_flutter.dart';
 import 'package:isar/isar.dart';
